chores: organised files into different folders
