+++ src/components/admin/semesters/SkeletonSemesterCard.tsx
@@ -1,4 +1,4 @@
-import { Skeleton } from "../ui/skeleton";
+import { Skeleton } from "../../ui/skeleton";
 
 export const SkeletonSemesterCard = () => {
   return (
