+++ src/components/ui/optionsPopover/OptionsButtonUtils.tsx
@@ -1,7 +1,7 @@
 import React, { forwardRef } from "react";
 import { SquarePen, Trash2 } from "lucide-react";
 
-import { Button } from "./ui/button";
+import { Button } from "../button";
 
 type OptionTypes = { type: "edit" | "delete" };
